Bug fix: multiple openocd instances can use the same vid/pid
If you have, on the same PC, multiple USB devices that use the mpsse
driver and that have the same vid/pid, there are common scenarios where
you can not use more than one.

For example:

Assume you have 2 USB devices with vid/pid/serial:

0x15ba/0x002b/S0001
0x15ba/0x002b/S0002

Assume that the USB devices are enumerated in that order by
libusb_get_device_list(...).

In your openocd_1.cfg, you request for the device with serial
number S0001.

openocd will correctly find this USB device, and open it using
libusb_open(...).

In your openocd_2.cfg, you request for the device with serial
number S0002.

When you start openocd with this second .cfg file, it will fail.

The reason is because the mpsse driver will call libusb_open(...) on the
first device with matching vid/pid. Since the first openocd instance
has already opened that device, this call will fail, LOG_ERROR
will be called, and openocd will abort.

The solution is to only call LOG_INFO, and move on to the
next USB device until the next device has been found.

An additional benefit of this patch is that it will also allow
using multiple mpsse devices with the same vid/pid and with the same
serial number.

E.g. the serial number on all Olimex ARM-USB-OCD-H devices is identical.

With this patch, for multiple openocd instances, the USB devices will
be assign in the order they are listed by libusb_get_device_list(...).